requirements_.txt
# Dependencias del proyecto corregido
# ====================================

# Framework web
Flask>=2.3.0
Werkzeug>=2.3.0

# Análisis de datos
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# APIs y HTTP
requests>=2.31.0
aiohttp>=3.8.0  # Para requests asíncronos

# Configuración de entorno
python-dotenv>=1.0.0

# Utilitarios
python-dateutil>=2.8.0
pathlib2>=2.3.0

# Caché y performance
cachetools>=5.3.0

# Desarrollo y testing (opcional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Sistema y logging
colorlog>=6.7.0